refactor: extract mobile sidebar components and improve production quality
- Extract UserProfileSection and RecentPastesSection from mobile sidebar for better maintainability and testability
- Add comprehensive type documentation explaining optionality patterns between components
- Optimize getUserInitials with useMemo to prevent unnecessary recalculations
- Improve production error handling with development-only console logging
- Enhance accessibility with proper ARIA labels for theme toggle and sign-out button
- Add loading animation to sign-out process for better UX feedback
