+++ src/features/dashboard/types/index.ts
@@ -14,14 +14,27 @@ export interface RecentPublicPaste {
   createdAt: Date;
 }
 
+// Mobile sidebar can function without paste data for basic user profile display
 export interface DashboardMobileSidebarProps {
   recentPublicPastes?: RecentPublicPaste[];
   totalPublicPastes?: number;
   user: User;
 }
 
+// Header component requires paste data for complete dashboard functionality  
 export interface DashboardHeaderProps {
   recentPublicPastes: RecentPublicPaste[];
   totalPublicPastes: number;
   user: User;
+}
+
+export interface UserProfileSectionProps {
+  user: User;
+  onClose: () => void;
+}
+
+export interface RecentPastesSectionProps {
+  recentPublicPastes: RecentPublicPaste[];
+  totalPublicPastes: number;
+  onClose: () => void;
 }
